requirements.txt
loguru~=0.7.3
pathlib~=1.0.1
SQLAlchemy~=2.0.41
python-telegram-bot~=22.0
pydantic-settings~=2.9.1
alembic~=1.15.2
pyTelegramBotAPI~=4.27.0
asyncpg~=0.30.0